refine code: Extract the code related to recall into a separate module, and remove all unused code
