+++ code/main/main.py
@@ -3,59 +3,38 @@ import logging
 import time
 import os
 import re
+import io
 from botocore import config
 from botocore.exceptions import ClientError,EventStreamError
 from datetime import datetime, timedelta
 import boto3
-import time
-import hashlib
 import uuid
 import base64
-# from transformers import AutoTokenizer
 from enum import Enum
-from opensearchpy import OpenSearch, RequestsHttpConnection
-from requests_aws4auth import AWS4Auth
-from langchain.chains.question_answering import load_qa_chain
 from langchain.chat_models import ChatOpenAI
-from langchain.embeddings import SagemakerEndpointEmbeddings
-from langchain.embeddings.sagemaker_endpoint import EmbeddingsContentHandler
-from langchain.llms.sagemaker_endpoint import LLMContentHandler,SagemakerEndpoint
 from langchain.prompts import PromptTemplate
-from langchain.chains import LLMChain,ConversationalRetrievalChain,ConversationChain
-from langchain.schema import BaseRetriever
-from langchain.schema import Document
-from langchain.llms.bedrock import Bedrock
 from pydantic import BaseModel,Field
-from langchain.pydantic_v1 import Extra, root_validator
 
 import openai
 from langchain.callbacks.base import BaseCallbackHandler
-from langchain.callbacks.manager import CallbackManagerForLLMRun
 from langchain_core.outputs import ChatGenerationChunk, GenerationChunk, LLMResult
 from typing import Any, Dict, List, Union,Mapping, Optional, TypeVar, Union
 from langchain.schema import LLMResult
-from langchain.llms.base import LLM
-import io
-import math
 from enum import Enum
 from boto3 import client as boto3_client
-from utils.web_search import web_search,add_webpage_content
 from utils.management import management_api,get_template
 from utils.utils import add_reference, render_answer_with_ref
 from generator.llm_wrapper import get_langchain_llm_model, invoke_model, format_to_message
+from retriever.hybrid_retriever import CustomDocRetriever
 
 lambda_client= boto3.client('lambda')
 dynamodb_client = boto3.resource('dynamodb')
-credentials = boto3.Session().get_credentials()
 region = boto3.Session().region_name
-awsauth = AWS4Auth(credentials.access_key, credentials.secret_key, region, 'es', session_token=credentials.token)
 
 DOC_INDEX_TABLE= 'chatbot_doc_index'
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
-sm_client = boto3.client("sagemaker-runtime")
 lambda_client = boto3_client('lambda')
-# llm_endpoint = 'bloomz-7b1-mt-2023-04-19-09-41-24-189-endpoint'
 chat_session_table = os.environ.get('chat_session_table')
 QA_SEP = "=>"
 A_Role="用户"
@@ -87,18 +66,10 @@ TOP_K = int(os.environ.get('TOP_K',4))
 NEIGHBORS = int(os.environ.get('neighbors',0))
 KNOWLEDGE_BASE_ID = os.environ.get('knowledge_base_id',None)
 
-BEDROCK_EMBEDDING_MODELID_LIST = ["cohere.embed-multilingual-v3","cohere.embed-english-v3","amazon.titan-embed-text-v1"]
 BEDROCK_LLM_MODELID_LIST = {'claude-instant':'anthropic.claude-instant-v1',
                             'claude-v2':'anthropic.claude-v2',
                             'claude-v3-sonnet': 'anthropic.claude-3-sonnet-20240229-v1:0',
                             'claude-v3-haiku' : 'anthropic.claude-3-haiku-20240307-v1:0'}
-
-boto3_bedrock = boto3.client(
-    service_name="bedrock-runtime",
-    region_name= os.environ.get('bedrock_region',region)
-)
-
-knowledgebase_client = boto3.client("bedrock-agent-runtime", region)
 
 ###记录跟踪日志，用于前端输出
 class TraceLogger(BaseModel):
@@ -197,339 +168,6 @@ class CustomStreamingOutCallbackHandler(BaseCallbackHandler):
         data = json.dumps({ 'msgid':self.msgid, 'role': "AI", 'text': {'content':str(error[0])+'[DONE]'},'connectionId':self.connectionId})
         self.postMessage(data)
 
-class ContentHandler(EmbeddingsContentHandler):
-    parameters = {
-        "max_new_tokens": 50,
-        "temperature": 0,
-        "min_length": 10,
-        "no_repeat_ngram_size": 2,
-    }
-    def transform_input(self, inputs: list[str], model_kwargs: Dict) -> bytes:
-        input_str = json.dumps({"inputs": inputs, **model_kwargs})
-        return input_str.encode('utf-8')
-
-    def transform_output(self, output: bytes) -> List[List[float]]:
-        response_json = json.loads(output.read().decode("utf-8"))
-        return response_json["sentence_embeddings"]
-
-class CustomDocRetriever(BaseRetriever):
-    embedding_model_endpoint :str
-    aos_endpoint: str
-    aos_index: str
-        
-    class Config:
-        """Configuration for this pydantic object."""
-        arbitrary_types_allowed = True
-        
-    @classmethod
-    def from_endpoints(cls,embedding_model_endpoint:str, aos_endpoint:str, aos_index:str,):
-        return cls(embedding_model_endpoint=embedding_model_endpoint,
-                  aos_endpoint=aos_endpoint,
-                  aos_index=aos_index)
-    
-    #this is for standard langchain interface
-    def get_relevant_documents(self, query_input: str) -> List[Document]:
-        recall_knowledge = None
-        if KNOWLEDGE_BASE_ID:
-            recall_knowledge = self.get_relevant_documents_from_bedrock(KNOWLEDGE_BASE_ID, query_input)
-        else:
-            recall_knowledge,_,_= self.get_relevant_documents_custom(query_input) 
-        top_k_results = []
-        for item in recall_knowledge:
-            top_k_results.append(Document(page_content=item.get('doc')))
-        return top_k_results
-       
-     ## kkn前置检索FAQ,，如果query非常相似，则返回作为cache
-    def knn_quick_prefetch(self,query_input: str) -> List[Any]:
-        global KNN_QUICK_PEFETCH_THRESHOLD
-        start = time.time()
-        query_embedding = get_vector_by_sm_endpoint(query_input, sm_client, self.embedding_model_endpoint)
-        elpase_time = time.time() - start
-        logger.info(f'knn_quick_prefetch, running time of get embeddings : {elpase_time:.3f}s')
-        aos_client = OpenSearch(
-                hosts=[{'host': self.aos_endpoint, 'port': 443}],
-                http_auth = awsauth,
-                use_ssl=True,
-                verify_certs=True,
-                connection_class=RequestsHttpConnection
-            )
-        start = time.time()
-        opensearch_knn_respose = search_using_aos_knn(aos_client,query_embedding[0], self.aos_index,size=3)
-        elpase_time = time.time() - start
-        logger.info(f'runing time of quick_knn_fetch : {elpase_time:.3f}s')
-        filter_knn_result = [item for item in opensearch_knn_respose if (item['score'] > KNN_QUICK_PEFETCH_THRESHOLD and item['doc_type'] == 'Question')]
-        if len(filter_knn_result) :
-            filter_knn_result.sort(key=lambda x:x['score'])
-        return filter_knn_result
-
-    async def aget_relevant_documents(self, query: str) -> List[Document]:
-        raise NotImplementedError
-    
-    
-    def add_neighbours_doc(self,client,opensearch_respose):
-        docs = []
-        docs_dict = {}
-        for item in opensearch_respose:
-            ## only apply to 'Paragraph','Sentence' type file
-            if item['doc_type'] in ['Paragraph','Sentence'] and ( 
-                item['doc_title'].endswith('.wiki.json') or 
-                item['doc_title'].endswith('.blog.json') or 
-                item['doc_title'].endswith('.txt') or 
-                item['doc_title'].endswith('.docx') or
-                item['doc_title'].endswith('.pdf')) :
-                #check if has duplicate content in Paragraph/Sentence types
-                key = f"{item['doc_title']}-{item['doc_category']}-{item['idx']}"
-                if key not in docs_dict:
-                    docs_dict[key] = item['idx']
-                    doc = self.search_paragraph_neighbours(client,item['idx'],item['doc_title'],item['doc_category'],item['doc_type'])
-                    docs.append({ **item, "doc": doc } )
-            else:
-                docs.append(item)
-        return docs
-
-    def search_paragraph_neighbours(self,client, idx, doc_title,doc_category,doc_type):
-        query ={
-            "query":{
-                "bool": {
-                "must": [
-                    {
-                    "terms": {
-                        "idx": [i for i in range(idx-NEIGHBORS,idx+NEIGHBORS+1)]
-                    }
-                    },
-                    {
-                    "terms": {
-                        "doc_title": [doc_title]
-                    }
-                    },
-                    {
-                  "terms": {
-                    "doc_category": [doc_category]
-                    }
-                    },
-                    {
-                    "terms": {
-                        "doc_type": [doc_type]
-                    }
-                    }
-                ]
-                }
-            }
-        }
-        query_response = client.search(
-            body=query,
-            index=self.aos_index
-        )
-         ## the 'Sentence' type has mappings sentence:content:idx = n:1:1, so need to filter out the duplicate idx
-        idx_dict = {}
-        doc = ''
-        for item in query_response["hits"]["hits"]:
-            key = item['_source']['idx']
-            if key not in idx_dict:
-                idx_dict[key]=key
-                doc += item['_source']['content']+'\n'            
-        return doc
-
-    ## 调用排序模型
-    def rerank(self,query_input: str, docs: List[Any],sm_client,cross_model_endpoint):
-        inputs = [query_input]*len(docs)
-        response_model = sm_client.invoke_endpoint(
-            EndpointName=cross_model_endpoint,
-            Body=json.dumps(
-                {
-                    "inputs": inputs,
-                    "docs": [item['doc'] for item in docs]
-                }
-            ),
-            ContentType="application/json",
-        )
-        json_str = response_model['Body'].read().decode('utf8')
-        json_obj = json.loads(json_str)
-        scores = json_obj['scores']
-        return scores if isinstance(scores, list) else [scores]
-    
-    def de_duplicate(self,docs):
-        unique_ids = set()
-        nodup = []
-        for item in docs:
-            doc_hash = hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest()
-            if doc_hash not in unique_ids:
-                nodup.append(item)
-                unique_ids.add(doc_hash)
-        return nodup
-
-    def get_relevant_documents_from_bedrock(self, knowledge_base_id:str, query_input:str):
-        BEDROCK_RETRIEVE_LIMIT=1000
-        response = knowledgebase_client.retrieve(
-            knowledgeBaseId=knowledge_base_id,
-            retrievalQuery={
-                'text': query_input[:BEDROCK_RETRIEVE_LIMIT]
-            },
-            retrievalConfiguration={
-                'vectorSearchConfiguration': {
-                    'numberOfResults': TOP_K,
-                    'overrideSearchType': 'HYBRID',
-                }
-            },
-        )
-
-        def remove_s3_prefix(s3_path):
-            return '/'.join(s3_path.split('/', 3)[3:])
-
-        all_docs = [ {'idx': 1, 'rank_score':0, 'doc_classify':'-', 'doc_author':'-', 'doc_category': '-', 'doc_type':'Paragraph', 'doc':item['content']['text'],'score':item['score'], 'doc_title': remove_s3_prefix(item['location']['s3Location']['uri']) } for item in response['retrievalResults']]
-        cross_model_endpoint = CROSS_MODEL_ENDPOINT
-        if cross_model_endpoint:
-            if all_docs:
-                scores = self.rerank(query_input, all_docs, sm_client, cross_model_endpoint)
-                sorted_indices = sorted(range(len(scores)), key=lambda i: scores[i], reverse=False)
-                recall_knowledge = [{**all_docs[idx],'rank_score':scores[idx] } for idx in sorted_indices[-TOP_K:] ]
-                recall_knowledge = [item for item in recall_knowledge if item['rank_score'] >= RERANK_THRESHOLD]
-                return recall_knowledge
-        
-        return all_docs
-    
-    def get_websearch_documents(self, query_input: str) -> list:
-        # 使用agent方式速度比较慢，直接改成调用search api
-        all_docs = web_search(query=query_input)
-        logger.info(f'all_docs:{all_docs}')
-        recall_knowledge = [{'doc_title':item['title'],'doc':item['title']+'\n'+item['snippet'],
-                             'doc_classify':'web_search','doc_type':'web_search','score':0.8,'doc_author':item['link']} for item in all_docs]
-        return recall_knowledge
-    
-    def get_relevant_documents_custom(self, query_input: str,use_search:bool = True):
-        global BM25_QD_THRESHOLD_HARD_REFUSE, BM25_QD_THRESHOLD_SOFT_REFUSE
-        global KNN_QQ_THRESHOLD_HARD_REFUSE, KNN_QQ_THRESHOLD_SOFT_REFUSE,WEBSEARCH_THRESHOLD
-        global KNN_QD_THRESHOLD_HARD_REFUSE, KNN_QD_THRESHOLD_SOFT_REFUSE
-        start = time.time()
-        query_embedding = get_vector_by_sm_endpoint(query_input, sm_client, self.embedding_model_endpoint)
-        elpase_time = time.time() - start
-        logger.info(f'running time of get embeddings : {elpase_time:.3f}s')
-        aos_client = OpenSearch(
-                hosts=[{'host': self.aos_endpoint, 'port': 443}],
-                http_auth = awsauth,
-                use_ssl=True,
-                verify_certs=True,
-                connection_class=RequestsHttpConnection
-            )
-        start = time.time()
-        opensearch_knn_respose = search_using_aos_knn(aos_client,query_embedding[0], self.aos_index,size=CHANNEL_RET_CNT)
-        elpase_time = time.time() - start
-        logger.info(f'runing time of opensearch_knn : {elpase_time}s seconds')
-        
-        # 4. get AOS invertedIndex recall
-        start = time.time()
-        opensearch_query_response = aos_search(aos_client, self.aos_index, "doc", query_input,size=CHANNEL_RET_CNT)
-        # logger.info(opensearch_query_response)
-        elpase_time = time.time() - start
-        logger.info(f'runing time of opensearch_query : {elpase_time}s seconds')
-
-        # 5. combine these two opensearch_knn_respose and opensearch_query_response
-        def combine_recalls(opensearch_knn_respose, opensearch_query_response):
-            '''
-            filter knn_result if the result don't appear in filter_inverted_result
-            '''
-            def get_topk_items(opensearch_knn_respose, opensearch_query_response, topk=1):
-
-                opensearch_knn_nodup = []
-                knn_unique_ids = set()
-                for item in opensearch_knn_respose:
-                    doc_hash = hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest()
-                    if doc_hash not in knn_unique_ids:
-                        opensearch_knn_nodup.append(item)
-                        knn_unique_ids.add(doc_hash)
-                
-                opensearch_bm25_nodup = []
-                bm25_unique_ids = set()
-                for item in opensearch_query_response:
-                    doc_hash = hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest()
-                    if doc_hash not in bm25_unique_ids:
-                        opensearch_bm25_nodup.append(item)
-                        doc_hash = hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest()
-                        bm25_unique_ids.add(doc_hash)
-
-                opensearch_knn_nodup.sort(key=lambda x: x['score'])
-                opensearch_bm25_nodup.sort(key=lambda x: x['score'])
-                logger.info(f'opensearch_knn_nodup count:{len(opensearch_knn_nodup)}')
-                logger.info(f'opensearch_bm25_nodup count:{len(opensearch_bm25_nodup)}')
-                
-                half_topk = math.ceil(topk/2) 
-    
-                kg_combine_result = [ item for item in opensearch_knn_nodup[-1*half_topk:]]
-
-                doc_hash = set([ hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest() for item in kg_combine_result ])
-
-                for item in opensearch_bm25_nodup:
-                    if len(kg_combine_result) >= topk:
-                        break
-                        
-                    if hashlib.md5(str(item['doc']).encode('utf-8')).hexdigest() not in doc_hash:
-                        kg_combine_result.append(item)
-                        logger.info(f'kg_combine_result.append')
-                    else:
-                        continue
-
-                return kg_combine_result
-            
-            ret_content = get_topk_items(opensearch_knn_respose, opensearch_query_response, TOP_K)
-            logger.info(f'get_topk_items:{len(ret_content)}')
-            return ret_content
-        
-        filter_knn_result = [ item for item in opensearch_knn_respose if item['score'] > KNN_QQ_THRESHOLD_HARD_REFUSE ]
-        filter_inverted_result = [ item for item in opensearch_query_response if item['score'] > BM25_QD_THRESHOLD_HARD_REFUSE ]
-
-        recall_knowledge = []
-        ##是否使用rerank
-        cross_model_endpoint = CROSS_MODEL_ENDPOINT
-        if cross_model_endpoint:
-            all_docs = filter_knn_result+filter_inverted_result
-
-            ###to do 去重
-            all_docs = self.de_duplicate(all_docs)
-            if all_docs:
-                scores = self.rerank(query_input, all_docs,sm_client,cross_model_endpoint)
-                ##sort by scores
-                sorted_indices = sorted(range(len(scores)), key=lambda i: scores[i], reverse=False)
-                recall_knowledge = [{**all_docs[idx],'rank_score':scores[idx] } for idx in sorted_indices[-TOP_K:] ] 
-                
-                ## 引入web search结果重新排序
-                if max(scores) < WEBSEARCH_THRESHOLD and use_search:
-                    web_knowledge = self.get_websearch_documents(query_input)
-                    if web_knowledge:
-                        search_scores = self.rerank(query_input, web_knowledge,sm_client,cross_model_endpoint)
-                        sorted_indices = sorted(range(len(search_scores)), key=lambda i: search_scores[i], reverse=False)
-                        
-                        ## 过滤websearch结果
-                        sorted_web_knowledge = [{**web_knowledge[idx],'rank_score':search_scores[idx] } for idx in sorted_indices if search_scores[idx]>=WEBSEARCH_THRESHOLD] 
-                        ## 前面返回的是snippet内容，可以对结果继续用爬虫抓取完整内容
-                        sorted_web_knowledge = add_webpage_content(sorted_web_knowledge)
-                        
-                        #添加到原有的知识里,并过滤到原来知识中的低分item
-                        recall_knowledge += sorted_web_knowledge
-                        #recall_knowledge = [item for item in  recall_knowledge if item['rank_score'] >= RERANK_THRESHOLD]
-            elif use_search:
-                ##如果没有找到知识，则直接搜索
-                web_knowledge = self.get_websearch_documents(query_input)
-                if web_knowledge:
-                    search_scores = self.rerank(query_input, web_knowledge,sm_client,cross_model_endpoint)
-                    sorted_indices = sorted(range(len(search_scores)), key=lambda i: search_scores[i], reverse=False)
-                    sorted_web_knowledge = [{**web_knowledge[idx],'rank_score':search_scores[idx] } for idx in sorted_indices if search_scores[idx]>=WEBSEARCH_THRESHOLD]
-                    ## 前面返回的是snippet内容，可以对结果继续用爬虫抓取完整内容
-                    recall_knowledge = add_webpage_content(sorted_web_knowledge)
-
-            #filter unrelevant knowledge by rerank score
-            recall_knowledge = [item for item in  recall_knowledge if item['rank_score'] >= RERANK_THRESHOLD]
-        else:
-            logger.info(f'filter_knn_result count:{len(filter_knn_result)}')
-            logger.info(f'filter_inverted_result count:{len(filter_inverted_result)}')
-            recall_knowledge = combine_recalls(filter_knn_result, filter_inverted_result)
-            recall_knowledge = [{**doc,'rank_score':0 } for doc in recall_knowledge]
-
-        ##如果是段落类型，添加临近doc
-        recall_knowledge = self.add_neighbours_doc(aos_client,recall_knowledge)
-
-        return recall_knowledge,opensearch_knn_respose,opensearch_query_response
-
-
 class ReplyStratgy(Enum):
     LLM_ONLY = 1
     WITH_LLM = 2
@@ -619,151 +257,7 @@ def chat_agent(query, detection):
 
     return answer,ref_doc
 
-def is_chinese(string):
-    for char in string:
-        if '\u4e00' <= char <= '\u9fff':
-            return True
 
-    return False
-
-
-def get_embedding_bedrock(texts,model_id):
-    provider = model_id.split(".")[0]
-    if provider == "cohere":
-        body = json.dumps({
-            "texts": [texts] if isinstance(texts, str) else texts,
-            "input_type": "search_document"
-        })
-    else:
-        # includes common provider == "amazon"
-        body = json.dumps({
-            "inputText": texts if isinstance(texts, str) else texts[0],
-        })
-    bedrock_resp = boto3_bedrock.invoke_model(
-            body=body,
-            modelId=model_id,
-            accept="application/json",
-            contentType="application/json"
-        )
-    response_body = json.loads(bedrock_resp.get('body').read())
-    if provider == "cohere":
-        embeddings = response_body['embeddings']
-    else:
-        embeddings = [response_body['embedding']]
-    return embeddings
-
-
-# AOS
-def get_vector_by_sm_endpoint(questions, sm_client, endpoint_name):
-    if endpoint_name in BEDROCK_EMBEDDING_MODELID_LIST:
-        return get_embedding_bedrock(questions,endpoint_name)
-
-    parameters = {
-    }
-
-    instruction_zh = "为这个句子生成表示以用于检索相关文章："
-    instruction_en = "Represent this sentence for searching relevant passages:"
-
-    if isinstance(questions, str):
-        instruction = instruction_zh if is_chinese(questions) else instruction_en
-    else:
-        instruction = instruction_zh
-
-    response_model = sm_client.invoke_endpoint(
-        EndpointName=endpoint_name,
-        Body=json.dumps(
-            {
-                "inputs": questions,
-                "parameters": parameters,
-                "is_query" : False,
-                "instruction" :  instruction
-            }
-        ),
-        ContentType="application/json",
-    )
-    json_str = response_model['Body'].read().decode('utf8')
-    json_obj = json.loads(json_str)
-    embeddings = json_obj['sentence_embeddings']
-    return embeddings
-
-def search_using_aos_knn(client, q_embedding, index, size=10):
-
-    #Note: 查询时无需指定排序方式，最临近的向量分数越高，做过归一化(0.0~1.0)
-    #精准Knn的查询语法参考 https://opensearch.org/docs/latest/search-plugins/knn/knn-score-script/
-    #模糊Knn的查询语法参考 https://opensearch.org/docs/latest/search-plugins/knn/approximate-knn/
-    #这里采用的是模糊查询
-    query = {
-        "size": size,
-        "query": {
-            "knn": {
-                "embedding": {
-                    "vector": q_embedding,
-                    "k": size
-                }
-            }
-        }
-    }
-    opensearch_knn_respose = []
-    query_response = client.search(
-        body=query,
-        index=index
-    )
-    opensearch_knn_respose = [{'idx':item['_source'].get('idx',1),'doc_category':item['_source']['doc_category'],'doc_classify':item['_source'].get('doc_classify'),'doc_title':item['_source']['doc_title'],'id':item['_id'],'doc':item['_source']['content'],"doc_type":item["_source"]["doc_type"],"score":item["_score"],'doc_author': item['_source']['doc_author'], 'doc_meta': item['_source'].get('doc_meta','')}  for item in query_response["hits"]["hits"]]
-    return opensearch_knn_respose
-    
-
-
-def aos_search(client, index_name, field, query_term, exactly_match=False, size=10):
-    """
-    search opensearch with query.
-    :param host: AOS endpoint
-    :param index_name: Target Index Name
-    :param field: search field
-    :param query_term: query term
-    :return: aos response json
-    """
-    if not isinstance(client, OpenSearch):   
-        client = OpenSearch(
-            hosts=[{'host': client, 'port': 443}],
-            http_auth = awsauth,
-            use_ssl=True,
-            verify_certs=True,
-            connection_class=RequestsHttpConnection
-        )
-    query = None
-    if exactly_match:
-        query =  {
-            "query" : {
-                "match_phrase":{
-                    "doc": {
-                        "query": query_term,
-                        "analyzer": "ik_smart"
-                      }
-                }
-            }
-        }
-    else:
-        query = {
-            "size": size,
-            "query": {
-                "match": { "content" : query_term }
-            },
-            "sort": [{
-                "_score": {
-                    "order": "desc"
-                }
-            }]
-        }
-    query_response = client.search(
-        body=query,
-        index=index_name
-    )
-
-    if exactly_match:
-        result_arr = [ {'idx':item['_source'].get('idx',0),'doc_category':item['_source']['doc_category'],'doc_classify':item['_source'].get('doc_classify'),'doc_title':item['_source']['doc_title'],'id':item['_id'],'doc': item['_source']['content'], 'doc_type': item['_source']['doc_type'], 'score': item['_score'],'doc_author': item['_source']['doc_author'], 'doc_meta': item['_source'].get('doc_meta','')} for item in query_response["hits"]["hits"]]
-    else:
-        result_arr = [ {'idx':item['_source'].get('idx',0),'doc_category':item['_source']['doc_category'],'doc_classify':item['_source'].get('doc_classify'),'doc_title':item['_source']['doc_title'],'id':item['_id'],'doc':item['_source']['content'], 'doc_type': item['_source']['doc_type'], 'score': item['_score'],'doc_author': item['_source']['doc_author'], 'doc_meta': item['_source'].get('doc_meta','')} for item in query_response["hits"]["hits"]]
-    return result_arr
 
 def delete_session(session_id,user_id):
     # dynamodb = boto3.resource('dynamodb')
@@ -798,9 +292,6 @@ def get_session(session_id,user_id):
     except Exception as e:
         logger.info(f"get session failed {str(e)}")
         return ""
-
-
-
 
 # param:    session_id
 #           question
@@ -948,31 +439,6 @@ Assistant:"""
         input_variables=["context",'question','chat_history', 'role_bot', 'ask_user_prompt']
     )
     return PROMPT
-
-# def create_qa_prompt_templete(prompt_template):
-#     if prompt_template == '':
-#         #prompt_template_zh = """{system_role_prompt} {role_bot}\n请根据反引号中的内容提取相关信息回答问题:\n```\n{chat_history}{context}\n```\n如果反引号中的内容为空,则回答不知道.\n用户:{question}"""
-#         prompt_template_zh = \
-# """{system_role_prompt}{role_bot}请根据以下的知识，回答用户的问题。
-# <context>
-# {context}
-# </context> 
-# 如果知识中的内容的包含markdown格式的内容，如参考图片，示意图，链接等，请尽可能利用并按markdown格式输出参考图片，示意图，链接。请严格基于跟问题相关的知识来回答问题，不要随意发挥和编造答案。请简洁有条理的回答，如果知识内容为空或者跟问题不相关，则回答不知道。
-# 前几轮的聊天记录如下，如果有需要请参考以下的记录。
-# <chat_history>
-# {chat_history} 
-# </chat_history>
-# Skip the preamble, go straight into the answer.
-# 用户问:{question} 
-# """
-#     else:
-#         prompt_template_zh = prompt_template
-#     PROMPT = PromptTemplate(
-#         template=prompt_template_zh,
-#         partial_variables={'system_role_prompt':SYSTEM_ROLE_PROMPT},
-#         input_variables=["context",'question','chat_history','role_bot']
-#     )
-#     return PROMPT
 
 def create_chat_prompt_templete(prompt_template='', llm_model_name='claude'):
     PROMPT = None
@@ -1168,7 +634,7 @@ def main_entry_new(user_id:str,wsconnection_id:str,session_id:str, query_input:s
     if use_qa:
         before_prefetch = time.time()
         TRACE_LOGGER.trace(f'**Prefetching cache...**')
-        cache_repsonses = doc_retriever.knn_quick_prefetch(query_input)
+        cache_repsonses = doc_retriever.knn_quick_prefetch(query_input=query_input, prefetch_threshold=KNN_QUICK_PEFETCH_THRESHOLD)
         elpase_time_cache = time.time() - before_prefetch
         TRACE_LOGGER.trace(f'**Running time of prefetching cache: {elpase_time_cache:.3f}s**')
         if cache_repsonses:
@@ -1266,10 +732,26 @@ def main_entry_new(user_id:str,wsconnection_id:str,session_id:str, query_input:s
         opensearch_query_response = []
         if KNOWLEDGE_BASE_ID:
             TRACE_LOGGER.trace('**Retrieving knowledge from bedrock knowledgebase...**')
-            recall_knowledge = doc_retriever.get_relevant_documents_from_bedrock(KNOWLEDGE_BASE_ID, query_input)
+            recall_knowledge = doc_retriever.get_relevant_documents_from_bedrock(
+                    knowledge_base_id=KNOWLEDGE_BASE_ID, 
+                    query_input=query_input, 
+                    top_k=TOP_K, 
+                    rerank_endpoint=CROSS_MODEL_ENDPOINT, 
+                    rerank_threshold=RERANK_THRESHOLD
+                )
         else:
             TRACE_LOGGER.trace('**Retrieving knowledge from OpenSearch...**')
-            recall_knowledge, opensearch_knn_respose, opensearch_query_response = doc_retriever.get_relevant_documents_custom(query_input,use_search) 
+            recall_knowledge, opensearch_knn_respose, opensearch_query_response = doc_retriever.get_relevant_documents_custom(
+                query_input=query_input, 
+                channel_return_cnt=CHANNEL_RET_CNT, 
+                top_k=TOP_K, 
+                knn_threshold=KNN_QQ_THRESHOLD_HARD_REFUSE, 
+                bm25_threshold=BM25_QD_THRESHOLD_HARD_REFUSE, 
+                web_search_threshold=WEBSEARCH_THRESHOLD, 
+                use_search=use_search,
+                rerank_endpoint=CROSS_MODEL_ENDPOINT,
+                rerank_threshold=RERANK_THRESHOLD
+            ) 
 
         elpase_time = time.time() - start
         logger.info(f'running time of opensearch_query : {elpase_time:.3f}s seconds')
@@ -1511,7 +993,17 @@ def lambda_handler(event, context):
         doc_retriever = CustomDocRetriever.from_endpoints(embedding_model_endpoint=embedding_endpoint,
                                     aos_endpoint= os.environ.get("aos_endpoint", ""),
                                     aos_index=aos_index)
-        recall_knowledge,opensearch_knn_respose,opensearch_query_response = doc_retriever.get_relevant_documents_custom(question) 
+        recall_knowledge,opensearch_knn_respose,opensearch_query_response = doc_retriever.get_relevant_documents_custom(
+                query_input=question, 
+                channel_return_cnt=CHANNEL_RET_CNT, 
+                top_k=TOP_K, 
+                knn_threshold=KNN_QQ_THRESHOLD_HARD_REFUSE, 
+                bm25_threshold=BM25_QD_THRESHOLD_HARD_REFUSE, 
+                web_search_threshold=WEBSEARCH_THRESHOLD, 
+                use_search=use_search,
+                rerank_endpoint=CROSS_MODEL_ENDPOINT,
+                rerank_threshold=RERANK_THRESHOLD
+            ) 
         extra_info = {"query_input": question, "opensearch_query_response" : opensearch_query_response, "opensearch_knn_respose": opensearch_knn_respose,"recall_knowledge":recall_knowledge }
         return {
             'statusCode': 200,
